% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performanceScores.R
\name{averagePrecision}
\alias{averagePrecision}
\title{Getting Average precision.}
\usage{
averagePrecision(uNumToExamineK, Yreal, Yhat)
}
\arguments{
\item{uNumToExamineK}{- The number of element of Yreal.}

\item{Yreal}{- A N x 1 vector that indicates that "1" is true, "0" is false.}

\item{Yhat}{- The estimated score(possibility) to be labeled as "1". The higher the element value is, The more likely to be labeled as "1".}
}
\value{
average precision
}
\description{
This function gives average precision.
Please refer to 'vignette' to find detailed concept of average precision.
}
\examples{
averagePrecision(2, c(1,0), c(0.5, 0.3))
averagePrecision(2, c(0,1), c(0.5, 0.3))
}
