% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataLoader.R
\name{dataLoader_yaml}
\alias{dataLoader_yaml}
\title{A function that convert a tree-shape dataset to the components for poincare-embeddings.}
\usage{
dataLoader_yaml(yamlDataset)
}
\arguments{
\item{yamlDataset}{- A tree-shape dataset with yaml format.}
}
\value{
entity - A vector of all the entities in the tree-shape dataset.

POS - A 2-row positive relation matrix that contains positive relation entities columnwisely.

NEG - A 2-row negative relation matrix that contains negative relation entities columnwisely.
}
\description{
A function that convert a tree-shape dataset to the components for poincare-embeddings.
}
\examples{
dataLoader_yaml(statistics)
}
