% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddingTrainer.R
\name{embeddingTrainer}
\alias{embeddingTrainer}
\title{Poincare-embedding trainer}
\usage{
embeddingTrainer(POS, NEG, entity, theta_dim = 2, N_epoch = 100,
  lr = 0.001, n_neg = 4)
}
\arguments{
\item{POS}{- A 2-row positive relation matrix that contains positive relation entities columnwisely.}

\item{NEG}{- A 2-row negative relation matrix that contains negative relation entities columnwisely.}

\item{entity}{- A vector of all the entities in the tree-shape dataset.}

\item{theta_dim}{- The dimension of the embedding space.}

\item{N_epoch}{- The number of epochs.}

\item{lr}{- The learning rate.}

\item{n_neg}{- The number of negative samples for each iteration.}
}
\value{
A trained matrix theta.
}
\description{
Poincare-embedding trainer
}
\examples{
statistics_yaml <- yaml::yaml.load(statistics)
statistics_tree <- data.tree::as.Node(statistics_yaml)
dataset <- dataLoader(statistics_tree)
embeddingTrainer(dataset$POS, dataset$NEG, dataset$entity, 2, 100, 0.001, 5)
}
