% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddingTrainer.R
\name{getDistanceGradVec}
\alias{getDistanceGradVec}
\title{Calculating Euclidean Gradient of distance function with respect to theta_i}
\usage{
getDistanceGradVec(theta_i, theta_j)
}
\arguments{
\item{theta_i}{- A N x 1 vector with N : Embedding space dimension. Each of theta works as proxy of an entity in tree-shape dataset.}

\item{theta_j}{- A N x 1 vector with N : Embedding space dimension. Each of theta works as proxy of an entity in tree-shape dataset.}
}
\value{
d(dist(theta_i, theta_j)) / d(theta_i) : A N x 1 vector.
}
\description{
This function calculates Euclidean Gradient of distance between theta_i and theta_j with respect to theta_i(the first argment of the function).
Please refer to 'vignette' to find the equation for Euclidean Gradient function and detailed ideas.
}
\examples{
getDistanceGradVec(c(0, 0), c(0.1, 0.5))
}
