% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddingTrainer.R
\name{getPoincareDistance}
\alias{getPoincareDistance}
\title{Poincare distance in matrix}
\usage{
getPoincareDistance(theta)
}
\arguments{
\item{theta}{- A N x M matrix with N : the number of entities and M : dimension of the embedding space.}
}
\value{
A N x N poincare distance matrix whose (i, j) element is poincare distance between theta_i and theta_j.
}
\description{
Poincare distance in matrix
}
\examples{
getPoincareDistance(matrix(stats::rnorm(100), 5, 20))
}
