% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/illustration.R
\name{make2dPlot}
\alias{make2dPlot}
\title{A 2-dimension plot for the poincare embeddings.}
\usage{
make2dPlot(theta, nameOfSample, POS, N_epoch, lr, n_neg)
}
\arguments{
\item{theta}{- A N x 2 matrix. Each of the column corresponds to the entity subject to embed.}

\item{nameOfSample}{- Name of the entity subject to embed.}

\item{POS}{- A 2-row positive relation matrix that contains positive relation entities columnwisely.}

\item{N_epoch}{- The number of epochs.}

\item{lr}{- The learning rate.}

\item{n_neg}{- The number of negative samples for each iteration.}
}
\value{
A 2-dimension plot with entity names and line segments between the entities with positive relation.
}
\description{
A 2-dimension plot for the poincare embeddings.
}
\examples{
toy_yaml <- yaml::yaml.load(toy)
toy_tree <- data.tree::as.Node(toy_yaml)
toy_dataset <- dataLoader(toy_tree)
emb <- poincareEmbeddings(toy_tree, 2, 100, 0.001, 5)
make2dPlot(emb$theta, toy_dataset$entity, toy_dataset$POS, 100, 0.001, 5)

}
