% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performanceScores.R
\name{performanceScores}
\alias{performanceScores}
\title{Getting Performance Scores : Ranking and Mean average precision.}
\usage{
performanceScores(distanceMtx, POS)
}
\arguments{
\item{distanceMtx}{- A N x N poincare distance matrix whose (i, j) element is poincare distance between theta_i and theta_j.}

\item{POS}{- A 2-row positive relation matrix that contains positive relation entities columnwisely.}
}
\value{
Ranking : The ranking of the true related entity.

MAP : Mean average precision.
}
\description{
This function gives mean ranking and mean average precision of the trained embedding entities, which are two main performance measures of graph model.
Please refer to 'vignette' to find detailed concepts of Ranking and Mean average precision.
}
\examples{
toy_yaml <- yaml::yaml.load(toy)
toy_tree <- data.tree::as.Node(toy_yaml)
toy_dataset <- dataLoader(toy_tree)
emb <- poincareEmbeddings(toy_tree, 2, 50, 0.001, 5)
distanceMtx <- getPoincareDistance(emb$theta)
performanceScores(distanceMtx, toy_dataset$POS)

}
