% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{poincareEmbeddings}
\alias{poincareEmbeddings}
\title{Title}
\usage{
poincareEmbeddings(yamlDataset, theta_dim = 2, N_epoch = 50,
  lr = 0.01, n_neg = 5)
}
\arguments{
\item{yamlDataset}{- A tree-shape dataset with yaml format.}

\item{theta_dim}{- The dimension of the embedding space.}

\item{N_epoch}{- The number of epochs.}

\item{lr}{- The learning rate.}

\item{n_neg}{- The number of negative samples for each iteration.}
}
\value{
Ranking : The ranking of the true related entity.

MAP : Mean average precision.

A 2-dimension plot only if theta_dim = 2.
}
\description{
Title
}
\examples{
poincareEmbeddings(statistics, 2, 100, 0.01, 5)
}
