% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddingTrainer.R
\name{proj}
\alias{proj}
\title{Projection function: to make thetas stay in the unit ball}
\usage{
proj(theta_i)
}
\arguments{
\item{-}{A N x 1 vector with N : Embedding space dimension. Each of theta works as proxy of an entity in tree-shape dataset.}
}
\value{
- A N x 1 vector inside the unit ball.
}
\description{
Projection function: to make thetas stay in the unit ball
}
\examples{
proj(c(2,2))
}
