% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddingTrainer.R
\name{getPoincareDistanceVec}
\alias{getPoincareDistanceVec}
\title{Calculating Poincare distance between two vectors.}
\usage{
getPoincareDistanceVec(theta_i, theta_j)
}
\arguments{
\item{theta_i}{- A N x 1 vector with N : Embedding space dimension. Each of theta works as proxy of an entity in tree-shape dataset.}

\item{theta_j}{- A N x 1 vector with N : Embedding space dimension. Each of theta works as proxy of an entity in tree-shape dataset.}
}
\value{
A poincare distance between theta_i and theta_j.
}
\description{
This function measures Poincare distance between two vectors.
Please refer to 'vignette' to find the equation for the Poincare distance and detailed ideas.
}
\examples{
getPoincareDistanceVec(c(0, 0), c(0.1, 0.5))
getPoincareDistanceVec(c(0.4, 0), c(0.1, 0.5))
}
