% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performanceScores.R
\name{performanceScores}
\alias{performanceScores}
\title{Title}
\usage{
performanceScores(distanceMtx, POS)
}
\arguments{
\item{distanceMtx}{- A N x N poincare distance matrix whose (i, j) element is poincare distance between theta_i and theta_j.}

\item{POS}{- A 2-row positive relation matrix that contains positive relation entities columnwisely.}
}
\value{
Ranking : The ranking of the true related entity.

MAP : Mean average precision.
}
\description{
Title
}
\examples{
performanceScores(distanceMtx, POS)
}
