% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{poincareEmbeddings}
\alias{poincareEmbeddings}
\title{Poincare-embeddings}
\usage{
poincareEmbeddings(dataset_tree, theta_dim = 2, N_epoch = 50,
  lr = 0.01, n_neg = 5)
}
\arguments{
\item{dataset_tree}{- A tree-shape dataset from "data.tree"}

\item{theta_dim}{- The dimension of the embedding space.}

\item{N_epoch}{- The number of epochs.}

\item{lr}{- The learning rate.}

\item{n_neg}{- The number of negative samples for each iteration.}
}
\value{
Ranking : The ranking of the true related entity.

MAP : Mean average precision.

A 2-dimension plot only if theta_dim = 2.
}
\description{
This function is aggregated the functions to give the outputs, 2-D plot(if theta_dim = 2), ranking and MAP.
Please refer to 'vignette' to find detailed implementation.
}
\examples{
statistics_adv_yaml <- yaml::yaml.load(statistics_adv)
statistics_adv_tree <- data.tree::as.Node(statistics_adv_yaml)
poincareEmbeddings(statistics_adv_tree, theta_dim = 2, N_epoch = 200, lr = 0.005, n_neg = 10)

}
