% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddingTrainer.R
\name{project}
\alias{project}
\title{Projection function.}
\usage{
project(theta_i)
}
\arguments{
\item{-}{A N x 1 vector with N : Embedding space dimension. Each of theta works as proxy of an entity in tree-shape dataset.}
}
\value{
- A N x 1 vector inside the unit ball.
}
\description{
This function make each embedding point theta stay in the unit ball by normalize the vector when Euclidean norm of the point is greater than 1.
Please refer to 'vignette' to find the equation for the projection function and detailed ideas.
}
\examples{
project(c(2,2))
}
